Crate rustc_next_trait_solver

Crate containing the implementation of the next-generation trait solver.

This crate may also contain things that are used by the old trait solver, but were uplifted in the process of making the new trait solver generic. So if you got to this crate from the old solver, it’s totally normal.

Modules

canonicalizer

coherence

delegate

placeholder

resolve

solve

The next-generation trait solver, currently still WIP.