rustc_next_trait_solver::canonicalizer
Struct Canonicalizer

pub struct Canonicalizer<'a, D: SolverDelegate<Interner = I>, I: Interner> {
    delegate: &'a D,
    canonicalize_mode: CanonicalizeMode,
    variables: &'a mut Vec<I::GenericArg>,
    primitive_var_infos: Vec<CanonicalVarInfo<I>>,
    variable_lookup_table: HashMap<I::GenericArg, usize>,
    binder_index: DebruijnIndex,
    cache: HashMap<(DebruijnIndex, I::Ty), I::Ty>,
}

Fields
delegate: &'a Dcanonicalize_mode: CanonicalizeModevariables: &'a mut Vec<I::GenericArg>primitive_var_infos: Vec<CanonicalVarInfo<I>>variable_lookup_table: HashMap<I::GenericArg, usize>binder_index: DebruijnIndexcache: HashMap<(DebruijnIndex, I::Ty), I::Ty>

We only use the debruijn index during lookup. We don’t need to track the variables as each generic arg only results in a single bound variable regardless of how many times it is encountered.

Implementations

impl<'a, D: SolverDelegate<Interner = I>, I: Interner> Canonicalizer<'a, D, I>

pub fn canonicalize<T: TypeFoldable<I>>( delegate: &'a D, canonicalize_mode: CanonicalizeMode, variables: &'a mut Vec<I::GenericArg>, value: T, ) -> Canonical<I, T>

fn get_or_insert_bound_var( &mut self, arg: impl Into<I::GenericArg>, canonical_var_info: CanonicalVarInfo<I>, ) -> BoundVar

fn finalize(self) -> (UniverseIndex, I::CanonicalVars)

fn cached_fold_ty(&mut self, t: I::Ty) -> I::Ty