rustc_next_trait_solver

Module coherence

Structs

OrphanChecker 🔒

UncoveredTyParams

Enums

Conflict

InCrate

Whether we do the orphan check relative to this crate or to some remote crate.

IsFirstInputType

OrphanCheckEarlyExit 🔒

OrphanCheckErr

OrphanCheckMode

Functions

orphan_check_trait_ref

Checks whether a trait-ref is potentially implementable by a crate.

trait_ref_is_knowable

Returns whether all impls which would apply to the trait_ref e.g. Ty: Trait<Arg> are already known in the local crate.

trait_ref_is_local_or_fundamental