Function rustc_next_trait_solver::solve::assembly::structural_traits::coroutine_closure_to_ambiguous_coroutine

fn coroutine_closure_to_ambiguous_coroutine<I: Interner>(
    cx: I,
    goal_kind: ClosureKind,
    goal_region: I::Region,
    def_id: I::DefId,
    args: CoroutineClosureArgs<I>,
    sig: CoroutineClosureSignature<I>,
) -> I::Ty

Given a coroutine-closure, project to its returned coroutine when we are not certain that the closure’s kind is compatible with the goal, and therefore also don’t know yet what the closure’s upvars are.

Note that we do not also push a AsyncFnKindHelper goal here.