rustc_next_trait_solver::solve

Module eval_ctxt

Modules

canonical 🔒

Canonicalization is used to separate some goal from its context, throwing away unnecessary information in the process.

probe 🔒

Structs

EvalCtxt

ReplaceAliasWithInfer 🔒

Eagerly replace aliases with inference variables, emitting AliasRelate goals, used when adding goals to the EvalCtxt. We compute the AliasRelate goals before evaluating the actual goal to get all the constraints we can.

Enums

CurrentGoalKind 🔒

The kind of goal we’re currently proving.

GenerateProofTree

Traits

SolverDelegateEvalExt