rustc_next_trait_solver::solve::eval_ctxt

Struct EvalCtxt

pub struct EvalCtxt<'a, D, I = I>
where
    D: SolverDelegate<Interner = I>,
    I: Interner,
{
    delegate: &'a D,
    variables: I::CanonicalVarKinds,
    current_goal_kind: CurrentGoalKind,
    pub(super) var_values: CanonicalVarValues<I>,
    pub(super) max_input_universe: UniverseIndex,
    pub(super) initial_opaque_types_storage_num_entries: <<D as SolverDelegate>::Infcx as InferCtxtLike>::OpaqueTypeStorageEntries,
    pub(super) search_graph: &'a mut SearchGraph<SearchGraphDelegate<D>>,
    nested_goals: Vec<(GoalSource, Goal<I, I::Predicate>, Option<GoalStalledOn<I>>)>,
    pub(super) origin_span: I::Span,
    tainted: Result<(), NoSolution>,
    pub(super) inspect: ProofTreeBuilder<D>,
}

Fields
delegate: &'a D

The inference context that backs (mostly) inference and placeholder terms instantiated while solving goals.

NOTE: The InferCtxt that backs the EvalCtxt is intentionally private, because the InferCtxt is much more general than EvalCtxt. Methods such as take_registered_region_obligations can mess up query responses, using At::normalize is totally wrong, calling evaluate_root_goal can cause coinductive unsoundness, etc.

Methods that are generally of use for trait solving are intentionally re-declared through the EvalCtxt below, often with cleaner signatures since we don’t care about things like ObligationCauses and Spans here. If some InferCtxt method is missing, please first think defensively about the method’s compatibility with this solver, or if an existing one does the job already.

variables: I::CanonicalVarKinds

The variable info for the var_values, only used to make an ambiguous response with no constraints.

current_goal_kind: CurrentGoalKind

What kind of goal we’re currently computing, see the enum definition for more info.

var_values: CanonicalVarValues<I>max_input_universe: UniverseIndex

The highest universe index nameable by the caller.

When we enter a new binder inside of the query we create new universes which the caller cannot name. We have to be careful with variables from these new universes when creating the query response.

Both because these new universes can prevent us from reaching a fixpoint if we have a coinductive cycle and because that’s the only way we can return new placeholders to the caller.

initial_opaque_types_storage_num_entries: <<D as SolverDelegate>::Infcx as InferCtxtLike>::OpaqueTypeStorageEntries

The opaque types from the canonical input. We only need to return opaque types which have been added to the storage while evaluating this goal.

search_graph: &'a mut SearchGraph<SearchGraphDelegate<D>>nested_goals: Vec<(GoalSource, Goal<I, I::Predicate>, Option<GoalStalledOn<I>>)>origin_span: I::Spantainted: Result<(), NoSolution>inspect: ProofTreeBuilder<D>
Implementations

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn compute_alias_relate_goal( &mut self, goal: Goal<I, (I::Term, I::Term, AliasRelationDirection)>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn assemble_and_evaluate_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, assemble_from: AssembleCandidatesFrom, ) -> Vec<Candidate<I>>

pub(super) fn forced_ambiguity( &mut self, cause: MaybeCause, ) -> Result<Candidate<I>, NoSolution>

fn assemble_impl_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, )

fn assemble_builtin_impl_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, )

fn assemble_param_env_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, )

fn assemble_alias_bound_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, )

fn assemble_alias_bound_candidates_recur<G: GoalKind<D>>( &mut self, self_ty: I::Ty, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, consider_self_bounds: AliasBoundKind, )

For some deeply nested <T>::A::B::C::D rigid associated type, we should explore the item bounds for all levels, since the associated_type_bounds feature means that a parent associated type may carry bounds for a nested associated type.

If we have a projection, check that its self type is a rigid projection. If so, continue searching by recursively calling after normalization.

fn assemble_object_bound_candidates<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, candidates: &mut Vec<Candidate<I>>, )

fn consider_coherence_unknowable_candidate<G: GoalKind<D>>( &mut self, goal: Goal<I, G>, ) -> Result<Candidate<I>, NoSolution>

In coherence we have to not only care about all impls we know about, but also consider impls which may get added in a downstream or sibling crate or which an upstream impl may add in a minor release.

To do so we return a single ambiguous candidate in case such an unknown impl could apply to the current goal.

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn filter_specialized_impls( &mut self, allow_inference_constraints: AllowInferenceConstraints, candidates: &mut Vec<Candidate<I>>, )

Check whether we can ignore impl candidates due to specialization.

This is only necessary for feature(specialization) and seems quite ugly.

pub(super) fn assemble_and_merge_candidates<G: GoalKind<D>>( &mut self, proven_via: Option<TraitGoalProvenVia>, goal: Goal<I, G>, inject_normalize_to_rigid_candidate: impl FnOnce(&mut EvalCtxt<'_, D>) -> QueryResult<I>, ) -> QueryResult<I>

Assemble and merge candidates for goals which are related to an underlying trait goal. Right now, this is normalizes-to and host effect goals.

We sadly can’t simply take all possible candidates for normalization goals and check whether they result in the same constraints. We want to make sure that trying to normalize an alias doesn’t result in constraints which aren’t otherwise required.

Most notably, when proving a trait goal by via a where-bound, we should not normalize via impls which have stricter region constraints than the where-bound:

trait Trait<'a> {
    type Assoc;
}

impl<'a, T: 'a> Trait<'a> for T {
    type Assoc = u32;
}

fn with_bound<'a, T: Trait<'a>>(_value: T::Assoc) {}

The where-bound of with_bound doesn’t specify the associated type, so we would only be able to normalize <T as Trait<'a>>::Assoc by using the impl. This impl adds a T: 'a bound however, which would result in a region error. Given that the user explicitly wrote that T: Trait<'a> holds, this is undesirable and we instead treat the alias as rigid.

See trait-system-refactor-initiative#124 for more details.

fn characterize_param_env_assumption( &mut self, param_env: I::ParamEnv, assumption: I::Clause, ) -> Result<CandidateSource<I>, NoSolution>

Compute whether a param-env assumption is global or non-global after normalizing it.

This is necessary because, for example, given:

where
    T: Trait<Assoc = u32>,
    i32: From<T::Assoc>,

The i32: From<T::Assoc> bound is non-global before normalization, but is global after. Since the old trait solver normalized param-envs eagerly, we want to emulate this behavior lazily.

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn compute_host_effect_goal( &mut self, goal: Goal<I, HostEffectPredicate<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn canonicalize_goal( &self, is_hir_typeck_root_goal: bool, goal: Goal<I, I::Predicate>, opaque_types: Vec<(OpaqueTypeKey<I>, I::Ty)>, ) -> (Vec<I::GenericArg>, CanonicalInput<I, I::Predicate>)

Canonicalizes the goal remembering the original values for each bound variable.

This expects goal and opaque_types to be eager resolved.

pub(in solve) fn evaluate_added_goals_and_make_canonical_response( &mut self, shallow_certainty: Certainty, ) -> QueryResult<I>

To return the constraints of a canonical query to the caller, we canonicalize:

var_values: a map from bound variables in the canonical goal to the values inferred while solving the instantiated goal.

external_constraints: additional constraints which aren’t expressible using simple unification of inference variables.

This takes the shallow_certainty which represents whether we’re confident that the final result of the current goal only depends on the nested goals.

In case this is Certainty::Maybe, there may still be additional nested goals or inference constraints required for this candidate to be hold. The candidate always requires all already added constraints and nested goals.

pub(in solve) fn make_ambiguous_response_no_constraints( &self, maybe_cause: MaybeCause, ) -> CanonicalResponse<I>

Constructs a totally unconstrained, ambiguous response to a goal.

Take care when using this, since often it’s useful to respond with ambiguity but return constrained variables to guide inference.

fn compute_external_query_constraints( &self, certainty: Certainty, normalization_nested_goals: NestedNormalizationGoals<I>, ) -> ExternalConstraintsData<I>

Computes the region constraints and new opaque types registered when proving a goal.

If an opaque was already constrained before proving this goal, then the external constraints do not need to record that opaque, since if it is further constrained by inference, that will be passed back in the var values.

pub(super) fn instantiate_and_apply_query_response( &mut self, param_env: I::ParamEnv, original_values: &[I::GenericArg], response: CanonicalResponse<I>, ) -> (NestedNormalizationGoals<I>, Certainty)

After calling a canonical query, we apply the constraints returned by the query using this function.

This happens in three steps:

we instantiate the bound variables of the query response

we unify the var_values of the response with the original_values

we apply the external_constraints returned by the query, returning the normalization_nested_goals

fn compute_query_response_instantiation_values<T: ResponseT<I>>( delegate: &D, original_values: &[I::GenericArg], response: &Canonical<I, T>, span: I::Span, ) -> CanonicalVarValues<I>

This returns the canonical variable values to instantiate the bound variables of the canonical response. This depends on the original_values for the bound variables.

fn unify_query_var_values( delegate: &D, param_env: I::ParamEnv, original_values: &[I::GenericArg], var_values: CanonicalVarValues<I>, span: I::Span, )

Unify the original_values with the var_values returned by the canonical query..

This assumes that this unification will always succeed. This is the case when applying a query response right away. However, calling a canonical query, doing any other kind of trait solving, and only then instantiating the result of the query can cause the instantiation to fail. This is not supported and we ICE in this case.

We always structurally instantiate aliases. Relating aliases needs to be different depending on whether the alias is rigid or not. We’re only really able to tell whether an alias is rigid by using the trait solver. When instantiating a response from the solver we assume that the solver correctly handled aliases and therefore always relate them structurally here.

fn register_region_constraints( &mut self, outlives: &[OutlivesPredicate<I, I::GenericArg>], )

fn register_new_opaque_types( &mut self, opaque_types: &[(OpaqueTypeKey<I>, I::Ty)], )

impl<'a, D, I> EvalCtxt<'a, D, I>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(in solve) fn probe<F, T>( &mut self, probe_kind: F, ) -> ProbeCtxt<'_, 'a, D, I, F, T>
where F: FnOnce(&T) -> ProbeKind<I>,

probe_kind is only called when proof tree building is enabled so it can be as expensive as necessary to output the desired information.

pub(in solve) fn probe_builtin_trait_candidate( &mut self, source: BuiltinImplSource, ) -> TraitProbeCtxt<'_, 'a, D, I, impl FnOnce(&QueryResult<I>) -> ProbeKind<I>>

pub(in solve) fn probe_trait_candidate( &mut self, source: CandidateSource<I>, ) -> TraitProbeCtxt<'_, 'a, D, I, impl FnOnce(&QueryResult<I>) -> ProbeKind<I>>

impl<'a, D, I> EvalCtxt<'a, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn typing_mode(&self) -> TypingMode<I>

pub(super) fn step_kind_for_source(&self, source: GoalSource) -> PathKind

Computes the PathKind for the step from the current goal to the nested goal required due to source.

See #136824 for a more detailed reasoning for this behavior. We consider cycles to be coinductive if they ‘step into’ a where-clause of a coinductive trait. We will likely extend this function in the future and will need to clearly document it in the rustc-dev-guide before stabilization.

pub(super) fn enter_root<R>( delegate: &D, root_depth: usize, generate_proof_tree: GenerateProofTree, origin_span: I::Span, f: impl FnOnce(&mut EvalCtxt<'_, D>) -> R, ) -> (R, Option<GoalEvaluation<I>>)

Creates a root evaluation context and search graph. This should only be used from outside of any evaluation, and other methods should be preferred over using this manually (such as SolverDelegateEvalExt::evaluate_root_goal).

pub(super) fn enter_canonical<R>( cx: I, search_graph: &'a mut SearchGraph<SearchGraphDelegate<D>>, canonical_input: CanonicalInput<I>, canonical_goal_evaluation: &mut ProofTreeBuilder<D>, f: impl FnOnce(&mut EvalCtxt<'_, D>, Goal<I, I::Predicate>) -> R, ) -> R

Creates a nested evaluation context that shares the same search graph as the one passed in. This is suitable for evaluation, granted that the search graph has had the nested goal recorded on its stack. This method only be used by search_graph::Delegate::compute_goal.

This function takes care of setting up the inference context, setting the anchor, and registering opaques from the canonicalized input.

fn evaluate_goal( &mut self, goal_evaluation_kind: GoalEvaluationKind, source: GoalSource, goal: Goal<I, I::Predicate>, stalled_on: Option<GoalStalledOn<I>>, ) -> Result<GoalEvaluation<I>, NoSolution>

Recursively evaluates goal, returning whether any inference vars have been constrained and the certainty of the result.

pub(super) fn evaluate_goal_raw( &mut self, goal_evaluation_kind: GoalEvaluationKind, source: GoalSource, goal: Goal<I, I::Predicate>, stalled_on: Option<GoalStalledOn<I>>, ) -> Result<(NestedNormalizationGoals<I>, GoalEvaluation<I>), NoSolution>

Recursively evaluates goal, returning the nested goals in case the nested goal is a NormalizesTo goal.

As all other goal kinds do not return any nested goals and NormalizesTo is only used by AliasRelate, all other callsites should use EvalCtxt::evaluate_goal which discards that empty storage.

pub(super) fn compute_goal( &mut self, goal: Goal<I, I::Predicate>, ) -> QueryResult<I>

pub(super) fn try_evaluate_added_goals( &mut self, ) -> Result<Certainty, NoSolution>

fn evaluate_added_goals_step(&mut self) -> Result<Option<Certainty>, NoSolution>

Iterate over all added goals: returning Ok(Some(_)) in case we can stop rerunning.

Goals for the next step get directly added to the nested goals of the EvalCtxt.

pub(crate) fn record_impl_args(&mut self, impl_args: I::GenericArgs)

Record impl args in the proof tree for later access by InspectCandidate.

pub(super) fn cx(&self) -> I

pub(super) fn add_goal( &mut self, source: GoalSource, goal: Goal<I, I::Predicate>, )

pub(super) fn add_goals( &mut self, source: GoalSource, goals: impl IntoIterator<Item = Goal<I, I::Predicate>>, )

pub(super) fn next_region_var(&mut self) -> I::Region

pub(super) fn next_ty_infer(&mut self) -> I::Ty

pub(super) fn next_const_infer(&mut self) -> I::Const

pub(super) fn next_term_infer_of_kind(&mut self, term: I::Term) -> I::Term

Returns a ty infer or a const infer depending on whether kind is a Ty or Const. If kind is an integer inference variable this will still return a ty infer var.

pub(super) fn term_is_fully_unconstrained( &self, goal: Goal<I, NormalizesTo<I>>, ) -> bool

Is the projection predicate is of the form exists<T> <Ty as Trait>::Assoc = T.

This is the case if the term does not occur in any other part of the predicate and is able to name all other placeholder and inference variables.

pub(super) fn eq<T: Relate<I>>( &mut self, param_env: I::ParamEnv, lhs: T, rhs: T, ) -> Result<(), NoSolution>

pub(super) fn relate_rigid_alias_non_alias( &mut self, param_env: I::ParamEnv, alias: AliasTerm<I>, variance: Variance, term: I::Term, ) -> Result<(), NoSolution>

This should be used when relating a rigid alias with another type.

Normally we emit a nested AliasRelate when equating an inference variable and an alias. This causes us to instead constrain the inference variable to the alias without emitting a nested alias relate goals.

pub(super) fn eq_structurally_relating_aliases<T: Relate<I>>( &mut self, param_env: I::ParamEnv, lhs: T, rhs: T, ) -> Result<(), NoSolution>

This sohuld only be used when we’re either instantiating a previously unconstrained “return value” or when we’re sure that all aliases in the types are rigid.

pub(super) fn sub<T: Relate<I>>( &mut self, param_env: I::ParamEnv, sub: T, sup: T, ) -> Result<(), NoSolution>

pub(super) fn relate<T: Relate<I>>( &mut self, param_env: I::ParamEnv, lhs: T, variance: Variance, rhs: T, ) -> Result<(), NoSolution>

pub(super) fn eq_and_get_goals<T: Relate<I>>( &self, param_env: I::ParamEnv, lhs: T, rhs: T, ) -> Result<Vec<Goal<I, I::Predicate>>, NoSolution>

Equates two values returning the nested goals without adding them to the nested goals of the EvalCtxt.

If possible, try using eq instead which automatically handles nested goals correctly.

pub(super) fn instantiate_binder_with_infer<T: TypeFoldable<I> + Copy>( &self, value: Binder<I, T>, ) -> T

pub(super) fn enter_forall<T: TypeFoldable<I>, U>( &mut self, value: Binder<I, T>, f: impl FnOnce(&mut Self, T) -> U, ) -> U

enter_forall, but takes &mut self and passes it back through the callback since it can’t be aliased during the call.

pub(super) fn resolve_vars_if_possible<T>(&self, value: T) -> T
where T: TypeFoldable<I>,

pub(super) fn eager_resolve_region(&self, r: I::Region) -> I::Region

pub(super) fn fresh_args_for_item(&mut self, def_id: I::DefId) -> I::GenericArgs

pub(super) fn register_ty_outlives(&self, ty: I::Ty, lt: I::Region)

pub(super) fn register_region_outlives(&self, a: I::Region, b: I::Region)

pub(super) fn well_formed_goals( &self, param_env: I::ParamEnv, term: I::Term, ) -> Option<Vec<Goal<I, I::Predicate>>>

Computes the list of goals required for arg to be well-formed

pub(super) fn trait_ref_is_knowable( &mut self, param_env: I::ParamEnv, trait_ref: TraitRef<I>, ) -> Result<bool, NoSolution>

pub(super) fn fetch_eligible_assoc_item( &self, goal_trait_ref: TraitRef<I>, trait_assoc_def_id: I::DefId, impl_def_id: I::DefId, ) -> Result<Option<I::DefId>, I::ErrorGuaranteed>

pub(super) fn register_hidden_type_in_storage( &mut self, opaque_type_key: OpaqueTypeKey<I>, hidden_ty: I::Ty, ) -> Option<I::Ty>

pub(super) fn add_item_bounds_for_hidden_type( &mut self, opaque_def_id: I::DefId, opaque_args: I::GenericArgs, param_env: I::ParamEnv, hidden_ty: I::Ty, )

pub(super) fn probe_existing_opaque_ty( &mut self, key: OpaqueTypeKey<I>, ) -> Option<(OpaqueTypeKey<I>, I::Ty)>

pub(super) fn evaluate_const( &self, param_env: I::ParamEnv, uv: UnevaluatedConst<I>, ) -> Option<I::Const>

pub(super) fn is_transmutable( &mut self, dst: I::Ty, src: I::Ty, assume: I::Const, ) -> Result<Certainty, NoSolution>

pub(super) fn replace_bound_vars<T: TypeFoldable<I>>( &self, t: T, universes: &mut Vec<Option<UniverseIndex>>, ) -> T

pub(super) fn may_use_unstable_feature( &self, param_env: I::ParamEnv, symbol: I::Symbol, ) -> bool

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn normalize_anon_const( &mut self, goal: Goal<I, NormalizesTo<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn normalize_free_alias( &mut self, goal: Goal<I, NormalizesTo<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn normalize_inherent_associated_term( &mut self, goal: Goal<I, NormalizesTo<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn normalize_opaque_type( &mut self, goal: Goal<I, NormalizesTo<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn compute_normalizes_to_goal( &mut self, goal: Goal<I, NormalizesTo<I>>, ) -> QueryResult<I>

pub fn instantiate_normalizes_to_term( &mut self, goal: Goal<I, NormalizesTo<I>>, term: I::Term, )

When normalizing an associated item, constrain the expected term to term.

We know term to always be a fully unconstrained inference variable, so eq should never fail here. However, in case term contains aliases, we emit nested AliasRelate goals to structurally normalize the alias.

pub fn structurally_instantiate_normalizes_to_term( &mut self, goal: Goal<I, NormalizesTo<I>>, term: AliasTerm<I>, )

Unlike instantiate_normalizes_to_term this instantiates the expected term with a rigid alias. Using this is pretty much always wrong.

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

fn translate_args( &mut self, goal: Goal<I, NormalizesTo<I>>, impl_def_id: I::DefId, impl_args: I::GenericArgs, impl_trait_ref: TraitRef<I>, target_container_def_id: I::DefId, ) -> Result<I::GenericArgs, NoSolution>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn compute_projection_goal( &mut self, goal: Goal<I, ProjectionPredicate<I>>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

fn consider_builtin_dyn_upcast_candidates( &mut self, goal: Goal<I, (I::Ty, I::Ty)>, a_data: I::BoundExistentialPredicates, a_region: I::Region, b_data: I::BoundExistentialPredicates, b_region: I::Region, ) -> Vec<Candidate<I>>

Trait upcasting allows for coercions between trait objects:

trait Super {}
trait Trait: Super {}
// results in builtin impls upcasting to a super trait
impl<'a, 'b: 'a> Unsize<dyn Super + 'a> for dyn Trait + 'b {}
// and impls removing auto trait bounds.
impl<'a, 'b: 'a> Unsize<dyn Trait + 'a> for dyn Trait + Send + 'b {}

fn consider_builtin_unsize_to_dyn_candidate( &mut self, goal: Goal<I, (I::Ty, I::Ty)>, b_data: I::BoundExistentialPredicates, b_region: I::Region, ) -> Result<Candidate<I>, NoSolution>

fn consider_builtin_upcast_to_principal( &mut self, goal: Goal<I, (I::Ty, I::Ty)>, source: CandidateSource<I>, a_data: I::BoundExistentialPredicates, a_region: I::Region, b_data: I::BoundExistentialPredicates, b_region: I::Region, upcast_principal: Option<Binder<I, ExistentialTraitRef<I>>>, ) -> Result<Candidate<I>, NoSolution>

fn consider_builtin_array_unsize( &mut self, goal: Goal<I, (I::Ty, I::Ty)>, a_elem_ty: I::Ty, b_elem_ty: I::Ty, ) -> Result<Candidate<I>, NoSolution>

We have the following builtin impls for arrays:

impl<T: ?Sized, const N: usize> Unsize<[T]> for [T; N] {}

While the impl itself could theoretically not be builtin, the actual unsizing behavior is builtin. Its also easier to make all impls of Unsize builtin as we’re able to use #[rustc_deny_explicit_impl] in this case.

fn consider_builtin_struct_unsize( &mut self, goal: Goal<I, (I::Ty, I::Ty)>, def: I::AdtDef, a_args: I::GenericArgs, b_args: I::GenericArgs, ) -> Result<Candidate<I>, NoSolution>

We generate a builtin Unsize impls for structs with generic parameters only mentioned by the last field.

struct Foo<T, U: ?Sized> {
    sized_field: Vec<T>,
    unsizable: Box<U>,
}
// results in the following builtin impl
impl<T: ?Sized, U: ?Sized, V: ?Sized> Unsize<Foo<T, V>> for Foo<T, U>
where
    Box<U>: Unsize<Box<V>>,
{}

fn disqualify_auto_trait_candidate_due_to_possible_impl( &mut self, goal: Goal<I, TraitPredicate<I>>, ) -> Option<Result<Candidate<I>, NoSolution>>

fn probe_and_evaluate_goal_for_constituent_tys( &mut self, source: CandidateSource<I>, goal: Goal<I, TraitPredicate<I>>, constituent_tys: impl Fn(&EvalCtxt<'_, D>, I::Ty) -> Result<Binder<I, Vec<I::Ty>>, NoSolution>, ) -> Result<Candidate<I>, NoSolution>

Convenience function for traits that are structural, i.e. that only have nested subgoals that only change the self type. Unlike other evaluate-like helpers, this does a probe, so it doesn’t need to be wrapped in one.

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

pub(super) fn unsound_prefer_builtin_dyn_impl( &mut self, candidates: &mut Vec<Candidate<I>>, )

FIXME(#57893): For backwards compatibility with the old trait solver implementation, we need to handle overlap between builtin and user-written impls for trait objects.

This overlap is unsound in general and something which we intend to fix separately. To avoid blocking the stabilization of the trait solver, we add this hack to avoid breakage in cases which are mostly fine™. Importantly, this preference is strictly weaker than the old behavior.

We only prefer builtin over user-written impls if there are no inference constraints. Importantly, we also only prefer the builtin impls for trait goals, and not during normalization. This means the only case where this special-case results in exploitable unsoundness should be lifetime dependent user-written impls.

pub(super) fn merge_trait_candidates( &mut self, candidates: Vec<Candidate<I>>, ) -> Result<(CanonicalResponse<I>, Option<TraitGoalProvenVia>), NoSolution>

pub(super) fn compute_trait_goal( &mut self, goal: Goal<I, TraitPredicate<I>>, ) -> Result<(CanonicalResponse<I>, Option<TraitGoalProvenVia>), NoSolution>

fn try_stall_coroutine( &mut self, self_ty: I::Ty, ) -> Option<Result<Candidate<I>, NoSolution>>

impl<'a, D, I> EvalCtxt<'a, D>
where D: SolverDelegate<Interner = I>, I: Interner,

fn compute_type_outlives_goal( &mut self, goal: Goal<I, OutlivesPredicate<I, I::Ty>>, ) -> QueryResult<I>

fn compute_region_outlives_goal( &mut self, goal: Goal<I, OutlivesPredicate<I, I::Region>>, ) -> QueryResult<I>

fn compute_coerce_goal( &mut self, goal: Goal<I, CoercePredicate<I>>, ) -> QueryResult<I>

fn compute_subtype_goal( &mut self, goal: Goal<I, SubtypePredicate<I>>, ) -> QueryResult<I>

fn compute_dyn_compatible_goal( &mut self, trait_def_id: I::DefId, ) -> QueryResult<I>

fn compute_well_formed_goal(&mut self, goal: Goal<I, I::Term>) -> QueryResult<I>

fn compute_unstable_feature_goal( &mut self, param_env: <I as Interner>::ParamEnv, symbol: <I as Interner>::Symbol, ) -> QueryResult<I>

fn compute_const_evaluatable_goal( &mut self, _: Goal<I, I::Const>, ) -> QueryResult<I>

fn compute_const_arg_has_type_goal( &mut self, goal: Goal<I, (I::Const, I::Ty)>, ) -> QueryResult<I>

impl<D, I> EvalCtxt<'_, D>
where D: SolverDelegate<Interner = I>, I: Interner,

fn try_merge_candidates( &mut self, candidates: &[Candidate<I>], ) -> Option<CanonicalResponse<I>>

Try to merge multiple possible ways to prove a goal, if that is not possible returns None.

In this case we tend to flounder and return ambiguity by calling [EvalCtxt::flounder].

fn bail_with_ambiguity( &mut self, candidates: &[Candidate<I>], ) -> CanonicalResponse<I>

fn flounder(&mut self, candidates: &[Candidate<I>]) -> QueryResult<I>

If we fail to merge responses we flounder and return overflow or ambiguity.

fn structurally_normalize_ty( &mut self, param_env: I::ParamEnv, ty: I::Ty, ) -> Result<I::Ty, NoSolution>

Normalize a type for when it is structurally matched on.

This function is necessary in nearly all cases before matching on a type. Not doing so is likely to be incomplete and therefore unsound during coherence.

fn structurally_normalize_const( &mut self, param_env: I::ParamEnv, ct: I::Const, ) -> Result<I::Const, NoSolution>

Normalize a const for when it is structurally matched on, or more likely when it needs .try_to_* called on it (e.g. to turn it into a usize).

This function is necessary in nearly all cases before matching on a const. Not doing so is likely to be incomplete and therefore unsound during coherence.

fn structurally_normalize_term( &mut self, param_env: I::ParamEnv, term: I::Term, ) -> Result<I::Term, NoSolution>

Normalize a term for when it is structurally matched on.

This function is necessary in nearly all cases before matching on a ty/const. Not doing so is likely to be incomplete and therefore unsound during coherence.

fn opaque_type_is_rigid(&self, def_id: I::DefId) -> bool