Struct rustc_next_trait_solver::solve::eval_ctxt::NestedGoals

pub struct NestedGoals<I: Interner> {
    pub normalizes_to_goals: Vec<Goal<I, NormalizesTo<I>>>,
    pub goals: Vec<(GoalSource, Goal<I, I::Predicate>)>,
}

Fields
normalizes_to_goals: Vec<Goal<I, NormalizesTo<I>>>

These normalizes-to goals are treated specially during the evaluation loop. In each iteration we take the RHS of the projection, replace it with a fresh inference variable, and only after evaluating that goal do we equate the fresh inference variable with the actual RHS of the predicate.

This is both to improve caching, and to avoid using the RHS of the projection predicate to influence the normalizes-to candidate we select.

Forgetting to replace the RHS with a fresh inference variable when we evaluate this goal results in an ICE..

goals: Vec<(GoalSource, Goal<I, I::Predicate>)>

The rest of the goals which have not yet processed or remain ambiguous.

Implementations

impl<I: Interner> NestedGoals<I>

pub fn new() -> Self

pub fn is_empty(&self) -> bool