rustc_next_trait_solver::solve::eval_ctxt
Struct ReplaceAliasWithInfer

struct ReplaceAliasWithInfer<'me, 'a, D, I>
where
    D: SolverDelegate<Interner = I>,
    I: Interner,
{
    ecx: &'me mut EvalCtxt<'a, D>,
    param_env: I::ParamEnv,
    cache: HashMap<I::Ty, I::Ty>,
}

Eagerly replace aliases with inference variables, emitting AliasRelate goals, used when adding goals to the EvalCtxt. We compute the AliasRelate goals before evaluating the actual goal to get all the constraints we can.

This is a performance optimization to more eagerly detect cycles during trait solving. See tests/ui/traits/next-solver/cycles/cycle-modulo-ambig-aliases.rs.

Fields
ecx: &'me mut EvalCtxt<'a, D>param_env: I::ParamEnvcache: HashMap<I::Ty, I::Ty>
Implementations

impl<'me, 'a, D, I> ReplaceAliasWithInfer<'me, 'a, D, I>
where D: SolverDelegate<Interner = I>, I: Interner,

fn new(ecx: &'me mut EvalCtxt<'a, D>, param_env: I::ParamEnv) -> Self