rustc_next_trait_solver::solve

Module trait_goals

Dealing with trait goals, i.e. T: Trait<'a, U>.

Enums

TraitGoalProvenVia 🔒

How we’ve proven this trait goal.

Functions

trait_predicate_with_def_id 🔒

Small helper function to change the def_id of a trait predicate - this is not normally something that you want to do, as different traits will require different args and so making it easy to change the trait is something of a footgun, but it is useful in the narrow circumstance of changing from MetaSized to Sized, which happens as part of the lazy elaboration of sizedness candidates.